+++ src/practices/index.ts
@@ -30,6 +30,10 @@ import { TimeToSolveIssuesPractice } from './LanguageIndependent/TimeToSolveIssu
 import { ThinPullRequestsPractice } from './LanguageIndependent/ThinPullRequestsPractice';
 import { SecurityVulnerabilitiesPractice } from './JavaScript/SecurityVulnerabilitiesPractice';
 import { FirstTestPractice, SecondTestPractice, InvalidTestPractice } from '../scanner';
+import { JavaPackageManagementUsedPractice } from './Java/JavaPackageManagementUsed';
+import { JavaLoggerUsedPractice } from './Java/JavaLoggerUsedPractice';
+import { JavaTestingFrameworkUsedPractice } from './Java/JavaTestingFrameworkUsedPractice';
+import { JavaMockingFrameworkUsedPractice } from './Java/JavaMockingFrameworkUsed';
 
 // register practices here
 export const practices = [
@@ -64,4 +68,8 @@ export const practices = [
   TimeToSolveIssuesPractice,
   ThinPullRequestsPractice,
   SecurityVulnerabilitiesPractice,
+  JavaPackageManagementUsedPractice,
+  JavaLoggerUsedPractice,
+  JavaTestingFrameworkUsedPractice,
+  JavaMockingFrameworkUsedPractice,
 ];